+++ source/Private/Assert-RequiredCommandParameter.ps1
@@ -32,6 +32,7 @@
 #>
 function Assert-RequiredCommandParameter
 {
+    [CmdletBinding()]
     param
     (
         [Parameter(Mandatory = $true)]